OSLC 3 implementation starting point
This is the starting point for developing the implementation of the
OSLC 3 capabilities. The service.js file has stubs for all the HTTP
functions that will be built using the ldp-service middleware component
for LDP capabilities.

These stubs need to be implemented to support the OSLC capabilities
such as discovery, resource preview, delegated dialogs, attachments,
query, etc.