Staring the re-implementation to decouple ldp-service
There is no longer a dependency on ldp-service because it was not practical to implement oslc-service on ldp-service using a simple chain of command. Rather oslc-service is build directly on storage services, implementing whatever part of LDP the service needs for its purposes. Servers that use oslc-service now configure the storage services for their needs.